pyproject.toml
[tool.poetry]
name = "thegamebot"
version = "2.0.0"
description = ""
authors = ["thegamecracks"]
license = "MPLv2.0"

[tool.poetry.dependencies]
python = "^3.10"
autopep8 = "^1.6.0"
dateparser = "^1.1.0"
emoji = "^1.6.3"
humanize = "^3.14.0"
inflect = "^5.4.0"
matplotlib = "^3.5.1"
mcstatus = "^6.5.0"
numpy = "^1.22.2"
Pillow = "^8.4.0"
Pint = "^0.18"
psutil = "^5.8.0"
python-dateutil = "^2.8.2"
python-dotenv = "^0.19.2"
qrcode = "^7.3.1"
asqlite = {git = "https://github.com/Rapptz/asqlite", rev = "ffc55a5b3d1b5e9c440c01620148be7fa6493ef5"}
abattlemetrics = {git = "https://github.com/thegamecracks/abattlemetrics", rev = "de7d75fbf306ee61279ac1503a24766cd23ceaaf"}
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
jishaku = {git = "https://github.com/Gorialis/jishaku"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
